Remove unreleased NetworkException
