+++ CHANGELOG.md
@@ -12,21 +12,18 @@ Please refer to [UPGRADING](UPGRADING.md) guide for upgrading to a major version
 - Added `cert_type` and `ssl_key_type` request options for TLS certificate and private-key file types
 - Added PHP stream handler support for the `ssl_key` request option
 - Added transport sharing via the `transport_sharing` client and cURL handler options
-- Added `GuzzleHttp\Exception\NetworkException` as the base class for network-related transfer failures
 
 ### Changed
 
 - Adjusted `guzzlehttp/psr7` version constraint to `^2.10`
 - Allowed domainless `SetCookie` instances to be stored without wildcard request matching
 - Changed no-proxy matching to respect request ports for host-and-port rules
-- Made `GuzzleHttp\Exception\ConnectException` extend `GuzzleHttp\Exception\NetworkException`
 - Prevented `CurlMultiHandler` destructors from throwing during cleanup
 - Improved invalid response handling across handlers
 
 ### Deprecated
 
 - Deprecated non-iterable `Pool` request collections, which will be rejected in 8.0
-- Deprecated `RequestException::getHandlerContext()` and `NetworkException::getHandlerContext()`, which will be removed in 8.0
 - Deprecated empty and malformed request protocol versions, which will be rejected in 8.0
 - Deprecated conflicting raw cURL request options, including `CURLOPT_SHARE`, which will be rejected in 8.0
 - Deprecated scalar-coerced `idn_conversion` request option values, which will be rejected in 8.0
