+++ tests/Exception/ConnectExceptionTest.php
@@ -3,7 +3,6 @@
 namespace GuzzleHttp\Tests\Exception;
 
 use GuzzleHttp\Exception\ConnectException;
-use GuzzleHttp\Exception\NetworkException;
 use GuzzleHttp\Psr7\Request;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Client\NetworkExceptionInterface;
@@ -18,12 +17,12 @@ class ConnectExceptionTest extends TestCase
     {
         $req = new Request('GET', '/');
         $prev = new \Exception();
-        $e = new ConnectException('foo', $req, $prev);
-        self::assertInstanceOf(NetworkException::class, $e);
+        $e = new ConnectException('foo', $req, $prev, ['foo' => 'bar']);
         self::assertInstanceOf(NetworkExceptionInterface::class, $e);
         self::assertNotInstanceOf(RequestExceptionInterface::class, $e);
         self::assertSame($req, $e->getRequest());
         self::assertSame('foo', $e->getMessage());
+        self::assertSame('bar', $e->getHandlerContext()['foo']);
         self::assertSame($prev, $e->getPrevious());
     }
 }
